feat(Money): 돈에 int 범위를 초과하는 숫자가 가능하도록 수정

+++ src/main/java/lotto/domain/LottoCreator.java
@@ -4,7 +4,7 @@ import lotto.constant.LottoConstants;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 final class LottoCreator {
     private static final String NULL_MESSAGE = "숫자를 뽑는 방식에 null 이 들어왔습니다";
@@ -22,9 +22,9 @@ final class LottoCreator {
     }
 
     public Lottos createLottos(Money money) {
-        int count = money.ableToBuy();
+        long count = money.ableToBuy();
         List<Lotto> lottos = new ArrayList<>();
-        IntStream.range(0, count)
+        LongStream.range(0, count)
                 .forEach(__ -> addLotto(lottos));
         return new Lottos(lottos);
     }
